+++ src/Render.hs
@@ -1,23 +1,7 @@
 module Render where
 
-import Game.Settings ( arena )
-import Game.Types
-    ( Game(Game),
-      GameState(Pause, Playing, GameOver),
-      Pos2D,
-      Snake(Snake, snake_body) )
-import Game.Utils ( squareSolid, i2f ) 
+import Game
 import Graphics.Gloss
-    ( blue,
-      red,
-      white,
-      yellow,
-      circle,
-      circleSolid,
-      color,
-      pictures,
-      translate,
-      Picture )
 import Data.Bifunctor ( Bifunctor(bimap) )
 
 drawSnake :: Snake -> [Picture]
@@ -27,7 +11,6 @@ drawSnake snake@(Snake (h:t) _ _) =
     translate (x*89 - 935) (y*69 - 550) (color yellow $ squareSolid 50)
     : drawSnake snake { snake_body = t }
 
-
 drawFruit :: Pos2D -> Picture
 drawFruit (x, y) =
   translate (x'*89 - 935) (y'*69 - 550)
@@ -36,12 +19,15 @@ drawFruit (x, y) =
         y'= i2f y
 
 drawGame :: Game -> IO Picture
-drawGame (Game snake fruit Playing) = return $
+drawGame (Game snake fruit Playing _) = return $
      pictures $
        drawFruit fruit : drawArena : drawSnake snake
 
-drawGame (Game _ _ GameOver) = return $ color white $ circleSolid 50
-drawGame (Game _ _ Pause)    = return $ circle 50
+drawGame (Game _ _ GameOver _) = return gameOverScreen
+drawGame (Game snake fruit Pause _)    = return $
+     pictures $ translate (-200) 0 (color white $ text "Pause") :
+       drawFruit fruit : drawArena : drawSnake snake
+
   -- pictures $ [ putIn (x, y) | (x, y) <- gameMap ]
   -- ++ drawSnake snake
   -- where putIn (x, y) =
@@ -51,6 +37,20 @@ drawGame (Game _ _ Pause)    = return $ circle 50
   --        where x'= int2Float x
   --              y'= int2Float y
 
+gameOverScreen :: Picture
+gameOverScreen =
+  playAgainTxt <> gameOverTxt
+
+gameOverTxt :: Picture
+gameOverTxt =
+  translate (-350) 0 $
+    color white $ text "Game Over"
+
+playAgainTxt :: Picture
+playAgainTxt =
+  scale 0.3 0.25 $
+    translate (-800) (-300) $
+      color white $ text "press SPC to play again"
 
 drawArena :: Picture
 drawArena = pictures [color red (squareSolid 50) `drawIn` (x, y) | (x, y) <- arena]