+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Revision history for my-snake-game
 
-## 0.1.0.0 -- YYYY-mm-dd
+## 0.1.0.0 -- 2025-06-27
 
 * First version. Released on an unsuspecting world.
+
+## 0.1.1.0 -- 2025-06-28 
+
+* Now the game has a "pause screen"
+and a "play again `"screen`"" too.